Moved LabelPainters to Label file
